separate files and components .. and delete unused files
