+++ src/components/UploadSection/cropImage.js
