+++ src/components/UploadSection/cropImage.stories.js
