+++ src/components/UploadSection/image.stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { OneImage } from './OneImage';
+
+export default {
+  title: 'sections/One Image Showcase',
+  component: OneImage
+};
+
+export const CropImageSection = () => {
+  return <OneImage />;
+};
