--- src/components/UploadSection/upload.stories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { UploadSection } from './';
-
-export default {
-  title: 'sections/upload section',
-  component: UploadSection
-};
-
-export const UploadSectionShwocase = () => {
-  return <UploadSection />;
-};
